Refactor/consolidation
This commit is comprised largely of a refactor and some code consolidation.

First, there were a handful of places where `namespace detail` material appeared amongst other, non-detail constructs, instead of in separate `detail.hpp` files. Now, all `detail::` constructs are defined in one or another `detail.hpp`.

I also did some refactoring of GNDStk's `convert` functions.

Broadly speaking, GNDStk has two major categories of functions called `convert`.

First, consider our "query system" that makes use of `Meta` and `Child` objects. With the Core Interface, a person can convert between user-defined types, and metadata and child nodes in our core tree structure. Metadata are stored as `std::strings` in the tree structure, while child nodes are stored as `Nodes`. Therefore, `convert` functions related to the query system include all of the following:

   - convert from `std::string` to user type
   - convert from user type to `std::string`
   - convert from `Node` to user type
   - convert from user type to `Node`

We already had several simple `convert` functions for the first two of these categories - for example, for converting between `string` and `int`. All of those have now been consolidated into the `convertQuery` directory. For the latter two we didn't have much, as we didn't see any well-defined built-in conversions to have between `Nodes`, which in general can contain arbitrary information, and any particular user type. The only exception was a `convert(Node,Node)`. For the sake of code consolidation, it has been moved to `Node.hpp`, which seems like a reasonable place.

The second broad category of `convert` functions, having an entirely different meaning than the ones described above, are the various conversions between our wrapper classes for various file formats - `XML`, `JSON`, and `HDF5` - and including `Tree` as well. Those were formerly in a `convert.hpp`, and files in the `convert/` directory. To more clearly distinguish these from the query-related `convert` functions, we've renamed those from `convert` to `convertClass`.

Essentially, code formerly in `type2string` and `string2type` locations now appears within `convertQuery`. What little was in `type2node` and `node2type` before (just the `convert(Node,Node)` as mentioned earlier) now appears directly in `Node.hpp`. Finally, `convert` for the class conversions, as opposed to the query-related conversions, is renamed to `convertClass`. The new system may or may not be perfect, but it's clearer than the old system was.

While doing the above work, I noticed that `detail::node2node` was actually called only from `Node`'s `operator=`. (At some time in the past, it may have been called from other places as well.) So, as a small simplification, I inlined it there, so that `detail::node2node` no longer needs to exist.

As a tiny but additional good deed, I removed end-of-line whitespace in every
file in which it occurred. All such files were in fact `CMakeLists.txt` files.
