+++ cmake/unit_testing.cmake
@@ -13,17 +13,14 @@ enable_testing()
 add_subdirectory( src/GNDStk/test )
 add_subdirectory( src/GNDStk/Defaulted/test )
 add_subdirectory( src/GNDStk/Tree/test )
-add_subdirectory( src/GNDStk/type2string/test )
-add_subdirectory( src/GNDStk/node2type/test )
-add_subdirectory( src/GNDStk/type2node/test )
-add_subdirectory( src/GNDStk/string2type/test )
 add_subdirectory( src/GNDStk/Meta/test )
 add_subdirectory( src/GNDStk/XML/test )
 add_subdirectory( src/GNDStk/JSON/test )
 add_subdirectory( src/GNDStk/HDF5/test )
 add_subdirectory( src/GNDStk/Node/test )
 add_subdirectory( src/GNDStk/Child/test )
-add_subdirectory( src/GNDStk/convert/test )
+add_subdirectory( src/GNDStk/convertQuery/test )
+add_subdirectory( src/GNDStk/convertClass/test )
 add_subdirectory( src/GNDStk/utility/test )
 add_subdirectory( src/GNDStk/Component/test )
 add_subdirectory( src/GNDStk/BlockData/test )
