+++ src/GNDStk/Meta/test/CMakeLists.txt
@@ -10,5 +10,5 @@ $<$<BOOL:${link_time_optimization}>:${${PREFIX}_link_time_optimization_flags}>
 $<$<BOOL:${nonportable_optimization}>:${${PREFIX}_nonportable_optimization_flags}>>
 
 ${CXX_appended_flags} ${GNDStk_appended_flags} )
-target_link_libraries( GNDStk.Meta.test PUBLIC GNDStk ) 
+target_link_libraries( GNDStk.Meta.test PUBLIC GNDStk )
 add_test( NAME GNDStk.Meta COMMAND GNDStk.Meta.test )
