+++ src/GNDStk/Node/src/assign.hpp
@@ -20,13 +20,21 @@ Node &operator=(Node &&other)
 Node &operator=(const Node &other)
 {
    if (&other != this) {
-      // The following uses add()s to add child nodes. The add()s,
-      // in turn, ensure that the parentNode pointers are correct.
-      detail::node2node(other,*this);
+      clear();
+      name = other.name;
+
+      // metadata
+      for (auto &m : other.metadata)
+         add(m.first,m.second);
+
+      // children
+      for (auto &c : other.children)
+         add() = *c;
 
       // validate
       for (auto &c : children)
          assert(c->parentNode = this);
    }
+
    return *this;
 }
