+++ src/GNDStk/Node/src/detail.hpp
@@ -1,16 +1,6 @@
 
 namespace detail {
 
-// node2node: forward declaration
-// This function is called by some of Node's assignment operators. We'd instead
-// put this forward declaration into the file in which *those* are defined, but
-// that file is #included inside class Node { ... }'s definition, where writing
-// the forward declaration wouldn't make sense.
-template<class NODE>
-void node2node(const NODE &, NODE &);
-
-
-
 // -----------------------------------------------------------------------------
 // isOptional
 // -----------------------------------------------------------------------------
