--- src/GNDStk/convert.hpp
@@ -1,8 +1,0 @@
-
-#include "GNDStk/convert/src/detail.hpp"
-
-// convert to:
-#include "GNDStk/convert/src/Tree.hpp"
-#include "GNDStk/convert/src/XML.hpp"
-#include "GNDStk/convert/src/JSON.hpp"
-#include "GNDStk/convert/src/HDF5.hpp"
