+++ src/GNDStk/convertClass.hpp
@@ -1,0 +1,8 @@
+
+#include "GNDStk/convertClass/src/detail.hpp"
+
+// convert to:
+#include "GNDStk/convertClass/src/Tree.hpp"
+#include "GNDStk/convertClass/src/XML.hpp"
+#include "GNDStk/convertClass/src/JSON.hpp"
+#include "GNDStk/convertClass/src/HDF5.hpp"
