+++ src/GNDStk/convertClass/test/CMakeLists.txt
@@ -1,6 +1,6 @@
 
-add_executable( GNDStk.type2string.test type2string.test.cpp )
-target_compile_options( GNDStk.type2string.test PRIVATE ${${PREFIX}_common_flags}
+add_executable( GNDStk.convertClass.test convertClass.test.cpp )
+target_compile_options( GNDStk.convertClass.test PRIVATE ${${PREFIX}_common_flags}
 $<$<BOOL:${strict}>:${${PREFIX}_strict_flags}>$<$<CONFIG:DEBUG>:
 ${${PREFIX}_DEBUG_flags}
 $<$<BOOL:${coverage}>:${${PREFIX}_coverage_flags}>>
@@ -10,5 +10,5 @@ $<$<BOOL:${link_time_optimization}>:${${PREFIX}_link_time_optimization_flags}>
 $<$<BOOL:${nonportable_optimization}>:${${PREFIX}_nonportable_optimization_flags}>>
 
 ${CXX_appended_flags} ${GNDStk_appended_flags} )
-target_link_libraries( GNDStk.type2string.test PUBLIC GNDStk ) 
-add_test( NAME GNDStk.type2string COMMAND GNDStk.type2string.test )
+target_link_libraries( GNDStk.convertClass.test PUBLIC GNDStk )
+add_test( NAME GNDStk.convertClass COMMAND GNDStk.convertClass.test )
