+++ src/GNDStk/convertQuery.hpp
@@ -1,0 +1,4 @@
+
+#include "GNDStk/convertQuery/src/string2type.hpp"
+#include "GNDStk/convertQuery/src/type2string.hpp"
+#include "GNDStk/convertQuery/src/detail.hpp"
