+++ src/GNDStk/convertQuery/src/detail.hpp
