+++ src/GNDStk/convertQuery/src/string2type.hpp
