+++ src/GNDStk/convertQuery/src/type2string.hpp
