+++ src/GNDStk/convertQuery/test/CMakeLists.txt
@@ -1,0 +1,19 @@
+
+add_executable( GNDStk.convertQuery.test
+                convertQuery.test.cpp
+                node2type.test.cpp
+                string2type.test.cpp
+                type2node.test.cpp
+                type2string.test.cpp )
+target_compile_options( GNDStk.convertQuery.test PRIVATE ${${PREFIX}_common_flags}
+$<$<BOOL:${strict}>:${${PREFIX}_strict_flags}>$<$<CONFIG:DEBUG>:
+${${PREFIX}_DEBUG_flags}
+$<$<BOOL:${coverage}>:${${PREFIX}_coverage_flags}>>
+$<$<CONFIG:RELEASE>:
+${${PREFIX}_RELEASE_flags}
+$<$<BOOL:${link_time_optimization}>:${${PREFIX}_link_time_optimization_flags}>
+$<$<BOOL:${nonportable_optimization}>:${${PREFIX}_nonportable_optimization_flags}>>
+
+${CXX_appended_flags} ${GNDStk_appended_flags} )
+target_link_libraries( GNDStk.convertQuery.test PUBLIC GNDStk )
+add_test( NAME GNDStk.convertQuery COMMAND GNDStk.convertQuery.test )
