+++ src/GNDStk/convertQuery/test/convertQuery.test.cpp
@@ -1,0 +1,8 @@
+
+#define CATCH_CONFIG_MAIN
+
+#include "catch.hpp"
+#include "GNDStk.hpp"
+
+SCENARIO("Testing convert() capabilities with respect to metadata") {
+}
