+++ src/GNDStk/convertQuery/test/node2type.test.cpp
@@ -1,6 +1,4 @@
 
-#define CATCH_CONFIG_MAIN
-
 #include "catch.hpp"
 #include "GNDStk.hpp"
